+++ src/turtle_food/include/turtle_food/turtle.h
@@ -35,12 +35,12 @@
 # include <ros/ros.h>
 # include <boost/shared_ptr.hpp>
 
-# include <turtlesim/Pose.h>
+# include <turtle_food/Pose.h>
 # include <geometry_msgs/Twist.h>
-# include <turtlesim/SetPen.h>
-# include <turtlesim/TeleportRelative.h>
-# include <turtlesim/TeleportAbsolute.h>
-# include <turtlesim/Color.h>
+# include <turtle_food/SetPen.h>
+# include <turtle_food/TeleportRelative.h>
+# include <turtle_food/TeleportAbsolute.h>
+# include <turtle_food/Color.h>
 #endif
 
 #include <QImage>
@@ -50,7 +50,7 @@
 
 #define PI 3.14159265
 
-namespace turtlesim
+namespace turtle_food
 {
 
 class Turtle
@@ -62,9 +62,9 @@ public:
   void paint(QPainter &painter);
 private:
   void velocityCallback(const geometry_msgs::Twist::ConstPtr& vel);
-  bool setPenCallback(turtlesim::SetPen::Request&, turtlesim::SetPen::Response&);
-  bool teleportRelativeCallback(turtlesim::TeleportRelative::Request&, turtlesim::TeleportRelative::Response&);
-  bool teleportAbsoluteCallback(turtlesim::TeleportAbsolute::Request&, turtlesim::TeleportAbsolute::Response&);
+  bool setPenCallback(turtle_food::SetPen::Request&, turtle_food::SetPen::Response&);
+  bool teleportRelativeCallback(turtle_food::TeleportRelative::Request&, turtle_food::TeleportRelative::Response&);
+  bool teleportAbsoluteCallback(turtle_food::TeleportAbsolute::Request&, turtle_food::TeleportAbsolute::Response&);
 
   void rotateImage();
