+++ src/turtle_food/include/turtle_food/turtle_frame.h
@@ -48,7 +48,7 @@
 # include "turtle.h"
 #endif
 
-namespace turtlesim
+namespace turtle_food
 {
 
 class TurtleFrame : public QFrame
